arquillian

+++ FullStackApplication/src/main/java/dao/CellDAO.java
@@ -9,6 +9,6 @@ import entities.CellHier;
 public interface CellDAO {
 
 	Collection<CellHier> getCell();
-	CellHier getByCellId(Long cellId);
-	void createCellHier(Long cellId, Long hier3_id, Long hier32_id, Long hier321_id);
+	CellHier getByCellId(Integer cellId);
+	void createCellHier(Integer cellId, Long hier3_id, Long hier32_id, Long hier321_id);
 }
